pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "satellite-desktop-upscaler"
version = "0.1.0"
description = "Cross-platform desktop app for satellite image upscaling."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.8,<7",
    "Pillow>=10,<13",
    "numpy>=1.26,<3",
    "rasterio>=1.3,<2",
    "affine>=2.4,<3",
    "platformdirs>=4.0,<5",
    "psutil>=5.9,<8",
]

[project.optional-dependencies]
dev = [
    "pytest>=9,<10",
    "pyinstaller>=6,<7",
]

[project.scripts]
satellite-upscale-cli = "backend.main:main"
satellite-upscale-gui = "app.ui:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []
